v1.1.1a
this commit is a continuation of the major 1.1.1 commit

the lower end of the tablet threshold gave too much dead space between two elements causing excess. that has been resolved by swapping the sizing of a container element and instead attributing the needed space to a padding value. 

the desktop style >= 1440 displays have a slight overhang on the x axis causing the browser to have a scroll bar, overflow is hidden in the X axis as the overhang is attributed to a element needing to overhang the 100 viewport width to achieve a smooth design